+++ src/components/Button/CancelButton.tsx
@@ -5,10 +5,10 @@ export const CancelButton = ({ ...props }) => {
 
   return (
     <button
+      className="btn"
       type="button"
-      className="flex h-8 items-center justify-center rounded bg-neutral-200 p-5 px-6 text-lg text-neutral-900 hover:bg-neutral-300"
-      {...props}
       onClick={() => router.push("/")}
+      {...props}
     >
       Cancel
     </button>