+++ tailwind.config.cjs
@@ -1,12 +1,9 @@
+/* eslint-disable global-require */
 /** @type {import('tailwindcss').Config} */
 module.exports = {
   content: ["./src/**/*.{js,ts,jsx,tsx}"],
   theme: {
     extend: {
-      colors: {
-        darkblue: "#00157b",
-        maroon: "#8c1a19",
-      },
       fontFamily: {
         sans: ["Prompt", "Helvetica", "Arial", "sans-serif"],
       },
@@ -15,5 +12,29 @@ module.exports = {
   variants: {
     borderWidth: ["last"],
   },
-  plugins: [require("@tailwindcss/forms")],
+  daisyui: {
+    themes: [
+      {
+        bcc: {
+          primary: "#1e40af",
+          "primary-content": "#ffffff",
+          secondary: "#831843",
+          "secondary-content": "#ffffff",
+          accent: "#374151",
+          "accent-content": "#ffffff",
+          neutral: "#e5e7eb",
+          "base-100": "#FFFFFF",
+          info: "#3b82f6",
+          "info-content": "#ffffff",
+          success: "#15803d",
+          "success-content": "#ffffff",
+          warning: "#FBBD23",
+          error: "#e53935",
+          "error-content": "#ffffff",
+        },
+      },
+      "dark",
+    ],
+  },
+  plugins: [require("@tailwindcss/forms"), require("daisyui")],
 };